Increased update speed to 240 updates per second

+++ ProjectRYB/src/com/RYB/Objects/Entity.java
@@ -3,10 +3,10 @@ package com.RYB.Objects;
 import java.awt.Graphics;
 
 public abstract class Entity {
-    public int x, y;
+    public float x, y;
     private int height, width;
     
-    public Entity(int x, int y, int width, int height){
+    public Entity(float x, float y, int width, int height){
         this.x = x;
         this.y = y;
         this.width = width;
